# Export: use Compositor to build images instead of pixel pushing (~4x faster at packing big metal/rough/occlusion textures)