~~Depends on #820.~~ merged

Makes more progress on #203.

-----

Whereas #820 created a happy path for encoding images, the unhappy path is still just as bad as before. This PR aims to improve the unhappy path.

It does this with the rather hacky method of creating a temp scene and using the Compositor to construct a node tree that will render the image described by the `ExportImage`. This is straight-forward to do from the description, for example, this

```
export_image.fills = {
    Channel.R: FillImage(image=..., src_chan=Channel.A),
    Channel.G: FillWhite(),
    Channel.B: FillImage(image=..., src_chan=Channel.G),
}
```

gives this Compositor graph

![compos](https://user-images.githubusercontent.com/11024420/69921921-de755680-145c-11ea-8058-ce31c9892646.png)

Then the scene is rendered to a tempfile and read back as usual.

## Color spaces

To use the Compositor, all images have their color-space set to `Non-Color` and the image is rendered with a `None` colorspace. I don't know if this is _correct_, but I _think_ this replicates what the pixel/numpy stuff was doing before. I've compared the PNGs this creates to master and found no differences (I also disabled the happy path and found no differences in the base color, etc. images either).

## Results

To test this, I made a material that intentionally goes down the unhappy path. The roughness channel is read from one channel of one generated image, and the metallic is read from one channel of another.

![example](https://user-images.githubusercontent.com/11024420/69921924-e59c6480-145c-11ea-8fa7-0f6fcc23a1c3.png)

Export times by the size of the generated textures (my RAM=4GB)

| Texture Size | Before | After |
|---|---|---|
| 1024x1024 | 0.96s | 0.24s |
| 4096x4096 | 16.6s | 3.7s |
| 8192x4096 | OOM | 7.2s |
| 8192x8192 | OOM | 15.9s |

As expected, the export time is roughly linear in the number of pixels to process. Using the compositor is about 4x faster than the code using `image.pixels` and numpy (ie. it works at about 4Mpixels/second vs 1Mpixels/second), at least for large images. Memory use also went down a little bit.

In practice, again, this only changes the unhappy path, so anything that already avoids it isn't helped. However the unhappy path can take an oversized portion of the export time. For example, this knocks another 50% off the export time for DamagedHelmet on top of the 50% that #820 already took.

| | master | #820 | This PR |
|---|---|---|---|
| DamagedHelmet.glb | 15s | 7s | 3.4s |

edit: #798 became #820.