@julienduroure 2019-12-04
Something that we need to be careful about: CombineRGBA does an automatic pre-multiply of alpha on  other channels (seems that this is something that we can't disable), so it can lead to wrong output.