@scurest 2019-12-04
Nice call! It looks like an Alpha Conversion from Straight to Premultiply (???) can fix it but I'd need to figure out why it works.

![alpha-premul](https://user-images.githubusercontent.com/11024420/70155156-57e89100-1677-11ea-9c93-4b352f1ef4c3.png)