@scurest 2019-12-04
Here's another problem: images you've painted on top of are _always_ in sRGB space and ignore their colorspace_settings, so their color comes out wrong

![painted-on-f](https://user-images.githubusercontent.com/11024420/70155461-ed842080-1677-11ea-8e1f-187f2a59281e.png)