@m0ppers 2019-12-06
OMG! thank you so much ❤️ great success here with your branch 🚀 Testing with 8k textures. Speedup is ENORMOUS and memory is finally acceptable