@emackey 2020-01-02
@scurest Is this obsolete now that #820 is merged?