@scurest 2020-01-03
@emackey Yeah, it was. I've rebased onto master though. It's ready for review now.

For the two problems discovered above, I just detect those cases and fallback to the numpy code. Fortunately the alpha problem doesn't affect metal-rough-occlusion textures at all, so you'll only hit it if your packing your textures and you're also painting on them in Blender—hopefully not the common case.

I considered adding a warning like `hit a slow path for image encoding; expect bad performance` in the numpy code for images greater than, say, 1Mpixel. But I don't know if it would be a good idea or not.